[BACKEND] API Validation - Auth Route - Moved validationRegisterInput and validateLoginInput functions from   validationMiddleware.js to validation/authValidation.js - Updated authRouter.js dependencies
